Decorator: stop runaway events; add unit tests (#126)

+++ .circleci/run_test_wrapper.sh
@@ -29,6 +29,14 @@ run_tests () {
 	EXIT_CODE=0
 
 	xvfb-run -a npm test || EXIT_CODE=$?
+	if [ -f /home/circleci/project/test_projects/proj0/prof.out ]; then
+		echo "[$0 ${FUNCNAME[0]}] copying profile output prof_${OE_VERSION}.out"
+		cp /home/circleci/project/test_projects/proj0/prof.out "/home/circleci/artifacts/prof_${OE_VERSION}.out"
+		if [ -f /home/circleci/project/test_projects/proj0/prof.json ]; then
+			echo "[$0 ${FUNCNAME[0]}] copying profile output prof_${OE_VERSION}.json"
+			cp /home/circleci/project/test_projects/proj0/prof.json "/home/circleci/artifacts/prof_${OE_VERSION}.json"
+		fi
+	fi
 	if [ "$EXIT_CODE" = "0" ]; then
 		echo "xvfb-run success"
 	else
@@ -59,6 +67,10 @@ save_and_print_debug_output () {
 
 run_lint () {
 	echo "[$0 ${FUNCNAME[0]}]"
+	if [ -n "${ABLUNIT_TEST_RUNNER_PROJECT_NAME:-}" ]; then
+		echo "[$0 ${FUNCNAME[0]}] skipping lint for single ABLUnit test runner project test"
+		return 0
+	fi
 
 	mkdir -p artifacts
 	rm -rf test_projects/proj7_load_performance/src/ADE-12.2.13.0
